Fix rubocop errors
