+++ classes/author.rb
