+++ classes/game.rb
