+++ modules/author_module.rb
