+++ modules/book_module.rb
