+++ modules/genre_module.rb
