+++ modules/label_module.rb
