+++ modules/load_module.rb
