+++ modules/music_module.rb
@@ -1,4 +1,4 @@
-require_relative '../classes/musicalbum'
+require_relative '../classes/music_album'
 require_relative '../classes/genre'
 require 'date'
