+++ spec/author_spec.rb
