+++ spec/game_spec.rb
